+++ src/main/java/by/exadel/internship/dto/time_for_call/TimeForCallDTO.java
@@ -1,10 +1,8 @@
-package by.exadel.internship.dto.timeForCall;
+package by.exadel.internship.dto.time_for_call;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import java.util.UUID;
 
 @Data
 @AllArgsConstructor
