+++ src/main/java/by/exadel/internship/service/SchedulingService.java
@@ -1,6 +1,6 @@
 package by.exadel.internship.service;
 
-import by.exadel.internship.dto.timeForCall.UserTimeSlotWithUserIdDTO;
+import by.exadel.internship.dto.time_for_call.UserTimeSlotWithUserIdDTO;
 import by.exadel.internship.dto.UserDTO;
 
 import java.util.List;
