+++ src/main/java/by/exadel/internship/service/TimeForCallService.java
@@ -1,6 +1,6 @@
 package by.exadel.internship.service;
 
-import by.exadel.internship.dto.timeForCall.TimeForCallDTO;
+import by.exadel.internship.dto.time_for_call.TimeForCallDTO;
 
 import java.util.List;
 import java.util.UUID;
