+++ src/main/java/by/exadel/internship/service/UserTimeSlotService.java
@@ -1,7 +1,7 @@
 package by.exadel.internship.service;
 
 
-import by.exadel.internship.dto.timeForCall.UserTimeSlotWithUserIdDTO;
+import by.exadel.internship.dto.time_for_call.UserTimeSlotWithUserIdDTO;
 import by.exadel.internship.dto.UserDTO;
 
 import java.util.UUID;
